login , member 수정

+++ src/main/java/com/mapapplication/mapapplication/controller/PlaceController.java
@@ -3,7 +3,9 @@ package com.mapapplication.mapapplication.controller;
 import com.mapapplication.mapapplication.dto.PlaceDto;
 import com.mapapplication.mapapplication.entity.Place;
 import com.mapapplication.mapapplication.entity.TripDailySchedule;
+import com.mapapplication.mapapplication.entity.TripSchedule;
 import com.mapapplication.mapapplication.repository.PlaceRepository;
+import com.mapapplication.mapapplication.repository.ScheduleRepository;
 import com.mapapplication.mapapplication.repository.TripDailyScheduleRepository;
 import com.mapapplication.mapapplication.service.PlaceService;
 import org.springframework.http.HttpStatus;
@@ -12,6 +14,7 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
 
 import javax.persistence.EntityNotFoundException;
+import javax.servlet.http.HttpSession;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -23,23 +26,26 @@ public class PlaceController {
     private final PlaceRepository placeRepository;
     private final PlaceService placeService;
     private final TripDailyScheduleRepository tripDailyScheduleRepository;
+    private final ScheduleRepository scheduleRepository;
 
     public PlaceController(PlaceRepository placeRepository, PlaceService placeService
-            , TripDailyScheduleRepository tripDailyScheduleRepository) {
+            , TripDailyScheduleRepository tripDailyScheduleRepository, ScheduleRepository scheduleRepository) {
         this.placeRepository = placeRepository;
         this.placeService = placeService;
         this.tripDailyScheduleRepository = tripDailyScheduleRepository;
+        this.scheduleRepository = scheduleRepository;
     }
 
-    @PostMapping("/save/{parentId}")
-    public String savePlaces(@PathVariable("parentId") Long parentId, @RequestBody List<PlaceDto> places) {
-        System.out.println("savePlaces 진입");
-        try {
+    @PostMapping("/{parentId}")
+    public String savePlaces(@PathVariable("parentId") Long parentId, @RequestBody List<PlaceDto> places,
+                             HttpSession session) {
+        Long userId = (Long) session.getAttribute("userId");
+        boolean isMatching = isUserIdMatching(userId, parentId);
+
+        if (isMatching) {
             placeService.savePlaces(parentId, places);
             return "redirect:/places/lists/" + parentId;  // 저장 성공 시 리다이렉트
-        } catch (EntityNotFoundException e) {
-            // Parent가 존재하지 않을 경우 처리할 예외 핸들링 로직
-            // 예: 에러 메시지 출력, 다른 경로로 리다이렉트 등
+        } else {
             return "error-page";
         }
     }
@@ -52,17 +58,24 @@ public class PlaceController {
      * @return
      */
     @GetMapping("/{parentId}")
-    public ModelAndView getPlacesByParentId(@PathVariable("parentId") Long parentId) {
-        List<Place> places = placeRepository.findByParentId(parentId);
+    public ModelAndView getPlacesByParentId(@PathVariable("parentId") Long parentId,
+                                            HttpSession session) {
+        ModelAndView modelAndView= new ModelAndView();
+        Long scheduleId = tripDailyScheduleRepository.findById(parentId).get().getParent().getId();
+        Long userId = (Long) session.getAttribute("userId");
+        boolean isMatching = isUserIdMatching(userId, scheduleId);
+        if (isMatching) {
+            List<Place> places = placeRepository.findByParentId(parentId);
+            System.out.println(places);
+            TripDailySchedule parent = tripDailyScheduleRepository.findById(parentId)
+                    .orElseThrow(() -> new EntityNotFoundException("부모 엔티티를 찾을 수 없음: " + parentId));
 
-        TripDailySchedule parent = tripDailyScheduleRepository.findById(parentId)
-                .orElseThrow(() -> new EntityNotFoundException("부모 엔티티를 찾을 수 없음: " + parentId));
-
-        ModelAndView modelAndView = new ModelAndView();
-        modelAndView.addObject("places", places);
-        modelAndView.addObject("daily", parent);
-        modelAndView.setViewName("map");
-
+            modelAndView.addObject("places", places);
+            modelAndView.addObject("daily", parent);
+            modelAndView.setViewName("map");
+        }else {
+            modelAndView.setViewName("error");
+        }
         return modelAndView;
     }
 
@@ -101,7 +114,8 @@ public class PlaceController {
 
     //장소 하나 지우기
     @DeleteMapping("/{id}")
-    public ResponseEntity<String> deletePlace(@PathVariable("id") Long id) {
+    public ResponseEntity<String> deletePlace(@PathVariable("id") Long id,
+                                              HttpSession session) {
         try {
             // 삭제할 장소 엔티티 가져오기
             Place place = placeRepository.findById(id)
@@ -118,17 +132,23 @@ public class PlaceController {
 
     // 부모id를 받아서 연관된 전체 장소 지워서 초기화
     @DeleteMapping("/reset/{parentId}")
-    public ResponseEntity<String> deletePlacesByParentId(@PathVariable("parentId") Long parentId) {
-        try {
+    public ResponseEntity<String> deletePlacesByParentId(@PathVariable("parentId") Long parentId,
+                                                         HttpSession session) {
+        Long userId = (Long) session.getAttribute("userId");
+        boolean isMatching = isUserIdMatching(userId, parentId);
+
+        if(isMatching) {
             placeRepository.deleteByParentId(parentId);
             return ResponseEntity.ok("places 삭제 완료");
-        } catch (Exception e) {
+        }else {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("places 삭제 실패");
         }
     }
 
-
+    public boolean isUserIdMatching(Long userId, Long tripScheduleId) {
+        TripSchedule tripSchedule = scheduleRepository.findById(tripScheduleId).orElse(null);
+        return ((tripSchedule != null) && tripSchedule.getUserId().equals(userId));
+    }
 
 
 }
-
